Projects/requirements.txt
# requirements.txt
streamlit
pandas
openpyxl
sentence-transformers
faiss-cpu
groq
python-dotenv